pin: forward interrupt note to the pinned proc
Prior to this patch we used the interrupt note to detach from a session.
This was only mildly convenient but left no nice way to interrupt the
pinned proc itself.  It also broke user's expectation that pressing the
interrupt (delete) key would kill the thing that appears to be running
in foreground.

We now ignore the interrupt note on the client side and instead forward
it to the pinned proc.  This matches the behaviour of programs running
inside the usual rio window context.