mq: implement 'coalesce' data mode
This implements the 'data coalesced' mode for the stream group.

Currently the order file is not useful in this mode since it only
communicates the names of streams which have some data in them.
For readers such as mq-cat this could mean moving the stream offset
past the global order presented by the order file, rendering it useless.

One solution that I can think of is exposing the length of the write in
the 'order' stream.  This would allow ordered reading of coalesced
streams.  However, naive reading would cancel out their only benefit.
The solution to /that/ is a reader who knows to combine the writes on
the same file that happened in sequence, therefore minimizing the number of read(2) calls needed.
An alternative solution is rewriting the 'order' stream on the server side.
This seems much harder to do right as we'd probably end up having to
maintain an independent order queue per reader.