mq: create an automatic root stream group
mq is quite often used in situations where only a single stream group
is needed.  In these cases it is preferable to just create the streams
immediately under the mq mount, avoiding the need for redundant naming
and an extra level in hierarchy.

We allow this by allocating an automatic root stream group at startup.