mq: fix Tcreate handling
When handling Tcreate we forgot that create(2) must also act as open(2),
that is, the server must initialize the state associated with the open
file in the same way it does when receiving an Topen call.

It also turned out that we needed to manually replace the fid->file
pointer with the new File, and also set the response qid to the new
file's qid.  This wasn't documented anywhere and it is likely that
the 9pfile(2) library can do it automatically, but until further
investigation we just do it ourselves.